requirements.txt
# === File: requirements.txt ===

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data processing
pandas==2.1.3
python-dotenv==1.0.0

# External APIs
facebook_business==19.0.0
openai>=1.6.1,<2.0.0

# Development
pydantic>=2.5.0,<3.0.0

# Rate limiting
slowapi==0.1.9
redis==5.0.1

# LangChain dependencies
langchain-core>=0.3.0,<0.4.0
langchain-azure-ai>=0.1.0,<0.2.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0